+++ ak_library_management/changelog.md
@@ -17,3 +17,12 @@ Changelog
 - Managed transient model in the wizard.
 - Improved the code quality by adding docstring.
 - Optimized code structure for better maintainability.
+
+[18.0.1.1.3] - 10-03-2025 | ADD Borrow Book transactions
+
+- Added the functions _cron_book_reminder_send_email and automated_action in borrow.transaction.history model.
+- Created schedule action,server action and automated action.
+- Created mail template.
+- Added comprehensive docstring to all the models and functions.
+- Improved the code quality by adding docstring.
+- Optimized code structure for better maintainability.